Improved compatibility between Handlebars.java and Handlebars.js.
Consider a scenaria such as this

---
{{#> test a=15}}
  {{#> test}}
    Hallo
  {{/test}}
{{/test}}

{{#*inline 'test'}}
  {{#if a}}
    {{a}}
  {{/if}}

  {{> @partial-block}}
{{/inline}}
---

In Handlebars.java "a" was known in both executions of "test". The extended context containing the parameter "a" was inherited by the call to @partial-block.
In Handlebars.js this context is not inherited and therefor "a" is only known during the evaluation of the outer "test", not in the inner "test".

This patch modifies the behavior of Handlebars.java to become more compatible in this regard...